--- hibernate-reactive/src/test/kotlin/com/linecorp/kotlinjdsl/test/reactive/HibernateStageReactiveQueryFactoryIntegrationTest.kt
@@ -1,195 +1,0 @@
-package com.linecorp.kotlinjdsl.test.reactive
-
-import com.linecorp.kotlinjdsl.listQuery
-import com.linecorp.kotlinjdsl.query.*
-import com.linecorp.kotlinjdsl.query.creator.SubqueryCreatorImpl
-import com.linecorp.kotlinjdsl.querydsl.expression.col
-import com.linecorp.kotlinjdsl.querydsl.from.associate
-import com.linecorp.kotlinjdsl.querydsl.from.fetch
-import com.linecorp.kotlinjdsl.singleQuery
-import com.linecorp.kotlinjdsl.test.WithKotlinJdslAssertions
-import com.linecorp.kotlinjdsl.test.entity.EntityDsl
-import com.linecorp.kotlinjdsl.test.entity.order.Order
-import com.linecorp.kotlinjdsl.test.entity.order.OrderGroup
-import com.linecorp.kotlinjdsl.test.entity.order.OrderItem
-import com.linecorp.kotlinjdsl.test.reactive.query.initFactory
-import com.linecorp.kotlinjdsl.updateQuery
-import kotlinx.coroutines.future.await
-import org.hibernate.reactive.stage.Stage.SessionFactory
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
-
-@ExtendWith(StageSessionFactoryExtension::class)
-internal class HibernateStageReactiveQueryFactoryIntegrationTest : EntityDsl, WithKotlinJdslAssertions {
-    private lateinit var sessionFactory: SessionFactory
-    private lateinit var queryFactory: HibernateStageReactiveQueryFactory
-
-    private val order1 = order { purchaserId = 1000 }
-    private val order2 = order { purchaserId = 1000 }
-    private val order3 = order { purchaserId = 1000 }
-    private val order4 = order { purchaserId = 2000 }
-
-    @BeforeEach
-    fun setUp() {
-        queryFactory = HibernateStageReactiveQueryFactory(
-            sessionFactory = sessionFactory, subqueryCreator = SubqueryCreatorImpl()
-        )
-        sequenceOf(order1, order2, order3, order4).forEach {
-            runBlocking {
-                retry(maxTries = 10, retryExceptions = listOf(NullPointerException::class, IllegalStateException::class)) {
-                    sessionFactory.withSession { session ->
-                        session.persist(it).thenCompose { session.flush() }
-                    }.await()
-                }
-            }
-        }
-    }
-
-    @Test
-    fun executeWithFactory() = runBlocking {
-        val order = order { purchaserId = 5000 }
-        val sessionFactory = initFactory()
-        retry(maxTries = 10, retryExceptions = listOf(NullPointerException::class, IllegalStateException::class)) {
-            val actual = sessionFactory.withSession { session ->
-                queryFactory.executeSessionWithFactory(session) { factory ->
-                    session.persist(order).thenCompose { session.flush() }.thenCompose {
-                        factory.singleQuery<Order> {
-                            select(entity(Order::class))
-                            from(entity(Order::class))
-                            where(col(Order::purchaserId).equal(5000))
-                        }
-                    }
-                }
-            }.await()
-
-            assertThat(actual.id).isEqualTo(order.id)
-        }
-        sessionFactory.close()
-    }
-
-    @Test
-    fun singleQuery() = runBlocking {
-        // when
-        val actual = queryFactory.singleQuery<Order> {
-            select(entity(Order::class))
-            from(entity(Order::class))
-            where(col(Order::purchaserId).equal(2000))
-        }
-
-        // then
-        assertThat(actual.id).isEqualTo(order4.id)
-    }
-
-    @Test
-    fun listQuery() = runBlocking {
-        // when
-        val actual = queryFactory.listQuery<Order> {
-            select(entity(Order::class))
-            from(entity(Order::class))
-            fetch(Order::groups)
-            fetch(OrderGroup::items)
-            fetch(OrderGroup::address)
-        }
-
-        // then
-        assertThat(actual).containsExactlyInAnyOrder(order1, order2, order3, order4)
-    }
-
-    @Test
-    fun subquery() = runBlocking {
-        val subquery = queryFactory.subquery<Long> {
-            val order = entity(Order::class, "o")
-
-            select(count(col(order, Order::id)))
-            from(order)
-            where(
-                col(order, Order::purchaserId).equal(col(Order::purchaserId)),
-            )
-        }
-
-        // then
-        assertThat(queryFactory.listQuery<Long> {
-            select(subquery)
-            from(entity(Order::class))
-            orderBy(col(Order::id).asc())
-        }).isEqualTo(listOf(order3.id, order3.id, order3.id, order1.id))
-    }
-
-
-    @Test
-    fun updateQuery() = runBlocking {
-        // when
-        queryFactory.updateQuery<Order> {
-            where(col(Order::purchaserId).equal(2000))
-            set(col(Order::purchaserId), 3000)
-        }
-
-        // then
-        val actual = queryFactory.singleQuery<Order> {
-            select(entity(Order::class))
-            from(entity(Order::class))
-            where(col(Order::purchaserId).equal(3000))
-        }
-        assertThat(actual.id).isEqualTo(order4.id)
-
-        assertThat(actual.purchaserId).isEqualTo(3000)
-    }
-
-    @Test
-    fun transaction() = runBlocking {
-        // when
-        try {
-            queryFactory.transactionWithFactory { queryFactory ->
-                queryFactory.listQuery<Order> {
-                    select(entity(Order::class))
-                    from(entity(Order::class))
-                    fetch(Order::groups)
-                    fetch(OrderGroup::items)
-                    fetch(OrderGroup::address)
-                }.thenCompose { orders ->
-                    queryFactory.updateQuery<Order> {
-                        where(col(Order::id).equal(orders.first().id))
-                        set(col(Order::purchaserId), orders.first().purchaserId + 1)
-                    }.executeUpdate.thenApply { orders }.thenCompose {
-                        queryFactory.updateQuery<Order> {
-                            throw IllegalStateException("transaction rollback")
-                        }.executeUpdate.thenApply { orders }
-                    }
-                }
-            }
-        } catch (e: IllegalStateException) {
-            assertThat(e).hasMessage("transaction rollback")
-        }
-
-        assertThat(queryFactory.transactionWithFactory { queryFactory ->
-            queryFactory.singleQuery<Order> {
-                select(entity(Order::class))
-                from(entity(Order::class))
-                fetch(Order::groups)
-                fetch(OrderGroup::items)
-                fetch(OrderGroup::address)
-                where(col(Order::id).equal(order1.id))
-            }
-        }).isEqualTo(order1)
-    }
-
-    @Test
-    fun deleteQuery() = runBlocking {
-        // when
-        queryFactory.deleteQuery<OrderItem> {
-            where(col(OrderItem::id).equal(order1.groups.first().items.first().id))
-        }
-
-        // then
-        val actual = queryFactory.listQuery<OrderItem> {
-            select(entity(OrderItem::class))
-            from(entity(OrderItem::class))
-            where(col(Order::id).equal(order1.id))
-            associate(OrderItem::group)
-            associate(OrderGroup::order)
-        }
-        assertThat(actual).isEmpty()
-    }
-}
-
