+++ hibernate-reactive/src/test/kotlin/com/linecorp/kotlinjdsl/test/reactive/criteriaquery/HibernateGroupByDslIntegrationTest.kt
@@ -2,8 +2,10 @@ package com.linecorp.kotlinjdsl.test.reactive.criteriaquery
 
 import com.linecorp.kotlinjdsl.test.reactive.HibernateCriteriaIntegrationTest
 import org.hibernate.reactive.stage.Stage
+import javax.persistence.EntityManagerFactory
 
 class HibernateGroupByDslIntegrationTest : HibernateCriteriaIntegrationTest,
     AbstractGroupByDslTest<Stage.SessionFactory>() {
     override lateinit var factory: Stage.SessionFactory
+    override lateinit var entityManagerFactory: EntityManagerFactory
 }
