+++ hibernate-reactive/src/test/kotlin/com/linecorp/kotlinjdsl/test/reactive/integration/querydsl/hint/HibernateCriteriaQueryDslHintIntegrationTest.kt
@@ -9,9 +9,11 @@ import com.linecorp.kotlinjdsl.test.reactive.runBlocking
 import org.hibernate.reactive.stage.Stage
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import javax.persistence.EntityManagerFactory
 
 internal class HibernateCriteriaQueryDslHintIntegrationTest : HibernateCriteriaIntegrationTest {
     override lateinit var factory: Stage.SessionFactory
+    override lateinit var entityManagerFactory: EntityManagerFactory
     private val order1 = order { purchaserId = 1000 }
     private val order2 = order { purchaserId = 1000 }
     private val order3 = order { purchaserId = 2000 }
